+++ p5/inc/symbolsTable.h
@@ -85,7 +85,7 @@ bool st_set_scalar_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  uint_fast16_t pos
+  int_fast32_t pos
 );
 
 /**
@@ -103,7 +103,7 @@ bool st_set_vector_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  uint_fast16_t pos,
+  int_fast32_t pos,
   uint_fast32_t size
 );
 
@@ -119,7 +119,7 @@ bool st_set_scalar_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  uint_fast16_t pos
+  int_fast32_t pos
 );
 
 /**
@@ -135,7 +135,7 @@ bool st_set_vector_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  uint_fast16_t pos,
+  int_fast32_t pos,
   uint_fast32_t size
 );
 
@@ -150,8 +150,8 @@ bool st_set_vector_parametre(
 bool st_set_function(
   SymbolsTable *st,
   String identifier,
-  uint_fast16_t params,
-  uint_fast16_t localvars
+  int_fast32_t params,
+  int_fast32_t localvars
 );
 
 /**
@@ -177,10 +177,10 @@ bool st_insertBlindCurrentScope(
   DataType dataType,
   IdentifierCategory identifierCategory,
   Scope scope,
-  uint_fast16_t pos,
+  int_fast32_t pos,
   uint_fast32_t size,
-  uint_fast16_t params,
-  uint_fast16_t localvars
+  int_fast32_t params,
+  int_fast32_t localvars
 );
 
 #endif
